pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Minimal URL shortener service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "aiogram>=3.19.0",
    "PyJWT>=2.9.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
